refactor: move the loop-prevention config beside its guards
RateLimitConfig, CircuitBreakerConfig and LoopPreventionConfig were read only by rate_limit.py, circuit_breaker.py and guard.py, all of which live in loop_prevention/. The top-level communication config now reaches them the way it already reaches conflict_resolution.config, and the module drops back under its size cap.
