+++ data/runtime_stats.yaml
@@ -1,6 +1,6 @@
 schema_version: 1
-last_generated_utc: '2026-08-11T00:38:30Z'
-generator_revision: 9ec14bb10
+last_generated_utc: '2026-08-11T00:47:50Z'
+generator_revision: 7bfbc0704
 stats:
   tests:
     raw: 45514
