+++ tests/integration/engine/test_multi_agent_delegation.py
@@ -15,15 +15,15 @@ from unittest.mock import AsyncMock
 
 import pytest
 
-from synthorg.communication.config import (
-    CircuitBreakerConfig,
-    HierarchyConfig,
-    LoopPreventionConfig,
-    RateLimitConfig,
-)
+from synthorg.communication.config import HierarchyConfig
 from synthorg.communication.delegation.authority import AuthorityValidator
 from synthorg.communication.delegation.hierarchy import HierarchyResolver
 from synthorg.communication.delegation.service import DelegationService
+from synthorg.communication.loop_prevention.config import (
+    CircuitBreakerConfig,
+    LoopPreventionConfig,
+    RateLimitConfig,
+)
 from synthorg.communication.loop_prevention.guard import DelegationGuard
 from synthorg.core.agent import AgentIdentity, ModelConfig, SkillSet
 from synthorg.core.company import Company, CompanyConfig
