+++ tests/unit/communication/delegation/test_service.py
@@ -4,12 +4,7 @@ from datetime import date
 
 import pytest
 
-from synthorg.communication.config import (
-    CircuitBreakerConfig,
-    HierarchyConfig,
-    LoopPreventionConfig,
-    RateLimitConfig,
-)
+from synthorg.communication.config import HierarchyConfig
 from synthorg.communication.delegation.authority import (
     AuthorityValidator,
 )
@@ -21,6 +16,11 @@ from synthorg.communication.delegation.record_store import (
 )
 from synthorg.communication.delegation.service import (
     DelegationService,
+)
+from synthorg.communication.loop_prevention.config import (
+    CircuitBreakerConfig,
+    LoopPreventionConfig,
+    RateLimitConfig,
 )
 from synthorg.communication.loop_prevention.guard import (
     DelegationGuard,
